fix: keep perf changes within source limits
